Small changes in notifications

+++ app/src/main/java/com/example/piston/main/notifications/NotificationsActivity.java
@@ -1,5 +1,9 @@
 package com.example.piston.main.notifications;
 
+import android.content.Context;
+import android.content.SharedPreferences;
+import android.content.res.Configuration;
+import android.os.Build;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -15,6 +19,7 @@ import androidx.lifecycle.ViewModelProvider;
 import com.example.piston.R;
 import com.example.piston.data.notifications.Notification;
 import com.example.piston.databinding.ActivityNotificationsBinding;
+import com.example.piston.utilities.Values;
 
 
 public class NotificationsActivity extends AppCompatActivity {
@@ -22,13 +27,35 @@ public class NotificationsActivity extends AppCompatActivity {
     private NotificationsAdapter adapter;
     ActionMode actionMode;
     ActionCallback actionCallback;
+    boolean darkModeEnabled;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_notifications);
         actionCallback = new ActionCallback();
-        adapter = new NotificationsAdapter(this);
+
+        SharedPreferences prefs = getSharedPreferences(Values.SHARED_PREFS, Context.MODE_PRIVATE);
+        boolean followSystem = false;
+
+        if (Build.VERSION.SDK_INT >= 29)
+            followSystem = prefs.getBoolean(Values.THEME_FOLLOW_SYSTEM, false);
+
+        if (followSystem) {
+            int nightModeFlags = getApplicationContext().getResources().getConfiguration().uiMode &
+                    Configuration.UI_MODE_NIGHT_MASK;
+            switch (nightModeFlags) {
+                case Configuration.UI_MODE_NIGHT_YES:
+                    darkModeEnabled = true;
+                    break;
+
+                case Configuration.UI_MODE_NIGHT_NO:
+                    darkModeEnabled = false;
+                    break;
+            }
+        }
+
+        adapter = new NotificationsAdapter(this, darkModeEnabled);
         NotificationsViewModel viewModel = new ViewModelProvider(this)
                 .get(NotificationsViewModel.class);
         ActivityNotificationsBinding binding = DataBindingUtil
@@ -69,8 +96,8 @@ public class NotificationsActivity extends AppCompatActivity {
 
         @Override
         public boolean onCreateActionMode(androidx.appcompat.view.ActionMode mode, Menu menu) {
-            NotificationMultiSelection.toggleStatusBarColor(NotificationsActivity.this, R.color.black);
-            mode.getMenuInflater().inflate(R.menu.notifications_top_app_bar, menu);
+            NotificationMultiSelection.toggleStatusBarColor(NotificationsActivity.this,
+                    (darkModeEnabled) ? R.color.design_default_color_secondary_variant : R.color.black);
             return true;
         }
 
@@ -81,15 +108,15 @@ public class NotificationsActivity extends AppCompatActivity {
 
         @Override
         public boolean onActionItemClicked(androidx.appcompat.view.ActionMode mode, MenuItem item) {
-            switch (item.getItemId()) {
-                case R.id.notification_delete:
-                    adapter.deleteNotifications();
-                    mode.finish();
-                    return true;
-                case R.id.notification_read:
-                    adapter.markAsRead();
-                    mode.finish();
-                    return true;
+            int itemId = item.getItemId();
+            if (itemId == R.id.notification_delete) {
+                adapter.deleteNotifications();
+                mode.finish();
+                return true;
+            } else if (itemId == R.id.notification_read) {
+                adapter.markAsRead();
+                mode.finish();
+                return true;
             }
             return false;
         }
@@ -98,8 +125,8 @@ public class NotificationsActivity extends AppCompatActivity {
         public void onDestroyActionMode(androidx.appcompat.view.ActionMode mode) {
             adapter.clearSelection();
             actionMode = null;
-            NotificationMultiSelection.toggleStatusBarColor(NotificationsActivity.this, R.color.design_default_color_primary_variant);
-
+            NotificationMultiSelection.toggleStatusBarColor(NotificationsActivity.this,
+                    (darkModeEnabled) ? R.color.design_default_color_primary_dark : R.color.black);
         }
     }
 }
